# backend/openApi.yaml
openapi: 3.0.0
info:
  title: SurfApp
  description: Web API to manage games and groups composed by them
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  version: 1.0.0-oas3
paths:
...
  /api/events:
    get:
      summary: fetches all the events
      operationId: getEvents
      responses:
        '200':
          description: response body should be a event object
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/event'
        '400':
          description: response body should be an error object
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/error'
        '401':
          description: response body should be an error object
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/error'
        '404':
          description: response body should be an error object
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/error'
        '500':
          description: response body should be an error object
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/error'
    post:
      summary: creates a event
      operationId: postEvent
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/api_events_body'
      responses:
        '201':
          description: response body should be a event object
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/id'
        '400':
          description: response body should be an error object
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/error'
        '401':
          description: response body should be an error object
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/error'
        '404':
          description: response body should be an error object
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/error'
        '409':
          description: response body should be an error object
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/error'
...
components:
  schemas:
...
    id:
      type: integer
...
    error:
      required:
        - message
      type: object
      properties:
        message:
          type: string
        status:
          type: integer
...
    event:
      type: object
      properties:
        id_:
          type: integer
        name_:
          type: string
        initial_date_:
          type: string
          format: date
        final_date_:
          type: string
          format: date
...
    api_events_body:
      type: object
      properties:
        name:
          type: string
        initial_date:
          type: string
          format: date
        final_date:
          type: string
          format: date
...